refactor: decompose bg-shell/index.ts into focused modules
Split the 3,179-line monolith into 7 focused modules:
- types.ts (251 lines): shared types, constants, pattern databases
- utilities.ts (55 lines): time formatting, Windows VT input restoration
- process-manager.ts (404 lines): process lifecycle, registry, persistence
- output-formatter.ts (259 lines): output analysis, digest, highlights
- readiness-detector.ts (126 lines): port probing, readiness detection
- interaction.ts (198 lines): send_and_wait, run on session, shell env query
- overlay.ts (432 lines): TUI process manager overlay

index.ts retains tool registration, command routing, footer, and event
handling (1,573 lines). All existing exports are preserved via re-exports.

Closes #519
